Address validation review findings
