+++ app/components/signup-form.tsx
@@ -5,7 +5,7 @@ import { useActionState } from "react";
 
 import { FieldErrorList } from "@/components/field-error-list";
 import { Link } from "@/i18n/navigation";
-import { authValidationLimits } from "@/lib/validation/auth-profile";
+import { authValidationLimits } from "@/lib/validation/auth-profile-limits";
 
 import { initialSignupActionState } from "../auth-action-state";
 import { signupAction } from "../auth-actions";
