+++ app/lib/validation/auth-profile-limits.ts
@@ -1,0 +1,8 @@
+export const authValidationLimits = {
+  displayNameMaxLength: 80,
+  emailMaxLength: 254,
+  passwordMaxLength: 128,
+  passwordMinLength: 8,
+  usernameMaxLength: 32,
+  usernameMinLength: 3,
+} as const;
