+++ app/lib/validation/auth-profile.ts
@@ -1,13 +1,6 @@
 import { z } from "zod";
 
-export const authValidationLimits = {
-  displayNameMaxLength: 80,
-  emailMaxLength: 254,
-  passwordMaxLength: 128,
-  passwordMinLength: 8,
-  usernameMaxLength: 32,
-  usernameMinLength: 3,
-} as const;
+import { authValidationLimits } from "./auth-profile-limits";
 
 const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 const usernamePattern = /^[A-Za-z0-9_-]+$/;
